move atlas dagger instantiation into its own project
