+++ atlasdb-cli/build.gradle
@@ -7,10 +7,8 @@ mainClassName = 'com.palantir.atlasdb.cli.AtlasCli'
 applicationName = 'atlasdb'
 
 dependencies {
-  compile project(':atlasdb-service')
-  compile project(':atlasdb-impl-shared')
+  compile project(':atlasdb-core')
   compile 'io.airlift:airline:0.7'
-  compile 'com.google.dagger:dagger:2.0.2'
 
   runtime project(':atlasdb-cassandra')
   runtime project(':atlasdb-rocksdb')
@@ -18,7 +16,6 @@ dependencies {
   testCompile project(':atlasdb-rocksdb')
 
   processor 'com.google.dagger:dagger-compiler:2.0.2'
-  processor 'org.immutables:value:' + libVersions.immutables
 }
 
 configurations.matching({ it.name in ['compile', 'runtime'] }).all {
