+++ atlasdb-cli/src/main/java/com/palantir/atlasdb/cli/command/timestamp/FastForwardTimestamp.java
@@ -18,7 +18,7 @@ package com.palantir.atlasdb.cli.command.timestamp;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import com.palantir.atlasdb.cli.services.AtlasDbServices;
+import com.palantir.atlasdb.services.AtlasDbServices;
 import com.palantir.timestamp.PersistentTimestampService;
 import com.palantir.timestamp.TimestampService;
