+++ atlasdb-cli/src/test/java/com/palantir/atlasdb/cli/command/TestSweepCommand.java
@@ -28,9 +28,7 @@ import com.google.common.collect.ImmutableSet;
 import com.google.common.io.BaseEncoding;
 import com.palantir.atlasdb.cli.runner.InMemoryTestRunner;
 import com.palantir.atlasdb.cli.runner.SingleBackendCliTestRunner;
-import com.palantir.atlasdb.cli.services.AtlasDbServicesFactory;
 import com.palantir.atlasdb.cli.services.DaggerTestAtlasDbServices;
-import com.palantir.atlasdb.cli.services.ServicesConfigModule;
 import com.palantir.atlasdb.cli.services.TestAtlasDbServices;
 import com.palantir.atlasdb.cli.services.TestSweeperModule;
 import com.palantir.atlasdb.keyvalue.api.Cell;
@@ -39,6 +37,8 @@ import com.palantir.atlasdb.keyvalue.api.Namespace;
 import com.palantir.atlasdb.keyvalue.api.TableReference;
 import com.palantir.atlasdb.keyvalue.api.Value;
 import com.palantir.atlasdb.protos.generated.TableMetadataPersistence;
+import com.palantir.atlasdb.services.AtlasDbServicesFactory;
+import com.palantir.atlasdb.services.ServicesConfigModule;
 import com.palantir.atlasdb.table.description.TableDefinition;
 import com.palantir.atlasdb.table.description.ValueType;
 import com.palantir.atlasdb.transaction.api.ConflictHandler;
