+++ atlasdb-cli/src/test/java/com/palantir/atlasdb/cli/command/TestTimestampCommand.java
@@ -47,10 +47,10 @@ import com.palantir.atlasdb.cleaner.SimplePuncher;
 import com.palantir.atlasdb.cli.command.timestamp.FetchTimestamp;
 import com.palantir.atlasdb.cli.runner.InMemoryTestRunner;
 import com.palantir.atlasdb.cli.runner.SingleBackendCliTestRunner;
-import com.palantir.atlasdb.cli.services.AtlasDbServicesFactory;
 import com.palantir.atlasdb.cli.services.DaggerTestAtlasDbServices;
-import com.palantir.atlasdb.cli.services.ServicesConfigModule;
 import com.palantir.atlasdb.cli.services.TestAtlasDbServices;
+import com.palantir.atlasdb.services.AtlasDbServicesFactory;
+import com.palantir.atlasdb.services.ServicesConfigModule;
 import com.palantir.common.time.Clock;
 import com.palantir.lock.LockClient;
 import com.palantir.lock.LockDescriptor;
