+++ atlasdb-cli/src/test/java/com/palantir/atlasdb/cli/services/TestAtlasDbServices.java
@@ -17,13 +17,18 @@ package com.palantir.atlasdb.cli.services;
 
 import javax.inject.Singleton;
 
+import com.palantir.atlasdb.services.AtlasDbServices;
+import com.palantir.atlasdb.services.KeyValueServiceModule;
+import com.palantir.atlasdb.services.LockAndTimestampModule;
+import com.palantir.atlasdb.services.RawKeyValueServiceModule;
+import com.palantir.atlasdb.services.ServicesConfigModule;
 import com.palantir.lock.LockClient;
 
 import dagger.Component;
 
 @Singleton
-@Component(modules = { ServicesConfigModule.class, KeyValueServiceModule.class, RawKeyValueServiceModule.class,
-        LockAndTimestampModule.class, TestSweeperModule.class, TestTransactionManagerModule.class })
+@Component(modules = {ServicesConfigModule.class, KeyValueServiceModule.class, RawKeyValueServiceModule.class,
+                      LockAndTimestampModule.class, TestSweeperModule.class, TestTransactionManagerModule.class })
 public abstract class TestAtlasDbServices extends AtlasDbServices {
 
     public abstract LockClient getTestLockClient();
