+++ atlasdb-core/build.gradle
@@ -1,0 +1,11 @@
+apply plugin: 'org.inferred.processors'
+
+apply from: '../gradle/shared.gradle'
+
+dependencies {
+  compile project(':atlasdb-server')
+  compile 'com.google.dagger:dagger:2.0.2'
+
+  processor 'com.google.dagger:dagger-compiler:2.0.2'
+  processor 'org.immutables:value:' + libVersions.immutables
+}
