+++ atlasdb-core/src/main/java/com/palantir/atlasdb/services/TransactionManagerModule.java
@@ -1,5 +1,5 @@
 /**
- * Copyright 2015 Palantir Technologies
+ * Copyright 2016 Palantir Technologies
  *
  * Licensed under the BSD-3 License (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.palantir.atlasdb.cli.services;
+package com.palantir.atlasdb.services;
 
 import javax.inject.Named;
 import javax.inject.Singleton;
