+++ atlasdb-core/src/test/java/com/palantir/atlasdb/services/AtlasDbServicesTest.java
@@ -1,0 +1,41 @@
+/**
+ * Copyright 2016 Palantir Technologies
+ *
+ * Licensed under the BSD-3 License (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://opensource.org/licenses/BSD-3-Clause
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.palantir.atlasdb.services;
+
+import java.io.File;
+import java.io.IOException;
+import java.net.URISyntaxException;
+import java.nio.file.Paths;
+
+import org.junit.Test;
+
+import com.palantir.atlasdb.config.AtlasDbConfigs;
+
+public class AtlasDbServicesTest {
+
+    @Test
+    public void daggerCanInstantiateAtlas() throws URISyntaxException, IOException {
+        File config = new File(getResourcePath("simple_atlas_config.yml"));
+        DaggerAtlasDbServices.builder()
+                .servicesConfigModule(ServicesConfigModule.create(AtlasDbConfigs.load(config)))
+                .build();
+    }
+
+    private static String getResourcePath(String fileName) throws URISyntaxException {
+        return Paths.get(AtlasDbServicesTest.class.getClassLoader().getResource(fileName).toURI()).toString();
+    }
+
+}
