+++ atlasdb-core/src/test/resources/simple_atlas_config.yml
@@ -1,0 +1,11 @@
+server:
+  applicationConnectors:
+    - type: http
+      port: 3828
+  adminConnectors:
+    - type: http
+      port: 3829
+
+atlasdb:
+  keyValueService:
+    type: memory
