+++ atlasdb-core/src/test/test.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+
+echo "test" >> ~/output
