+++ settings.gradle
@@ -7,6 +7,7 @@ include ":atlasdb-client"
 include ":atlasdb-client-protobufs"
 include ":atlasdb-config"
 include ":atlasdb-console"
+include ":atlasdb-core"
 include ":atlasdb-dropwizard-bundle"
 include ":atlasdb-dropwizard-tests"
 include ":atlasdb-dbkvs"
